refactored generation into independent functions, fixed collection+finalmap off by one
when calculating the number of slots needed for collecting we weren't taking into
account that if the current nextII was %64=0 we'd actually be adding two since we
have to account for the finalmap chunks threaded through the indirections. made
incrementing into a helper function to make it more explicit

still searching for an error that causes finalize() calls when no objects require
them. it's not many calls. probably an off by one.